% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{new_check}
\alias{new_check}
\title{Generate a function to check an object}
\usage{
new_check(test, error_args, env = rlang::caller_env())
}
\arguments{
\item{test}{`[expression]`

An expression in terms of `x` which returns `TRUE` or `FALSE`, used to test
some quality of an object `x`.

Examples of tests include:
- does `x` contain any `NaN` values, `any(is.nan(x))`
- is `x` a date before the 21st century, `is.Date(x) && x < as.Date("2000-01-01")`
- is `x` an even number, `is.numeric(x) && x %% 2 == 0`}

\item{error_args}{`[character / list / expression]`

Arguments passed to [cli::cli_abort()] by the generated function when `test`
evaluates to `FALSE`.}

\item{env}{`[environment]`

The environment of the generated function. By default, `env` is the current
environment.}
}
\value{
A function with arguments `x`, `x_name`, `error_call`, and `error_class`,
which returns `x` if `x` meets some criteria (specified by `test`) and
which raises an error otherwise.
}
\description{
[new_check()], [new_check_must()], and [new_check_must_not()] generate a
function which checks an object using a specified `test`. If the `test`
evaluates to `FALSE`, an error is raised. Otherwise, the checked object is
returned.
}
