% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyr-package.R
\docType{package}
\name{specifyr-package}
\alias{specifyr}
\alias{specifyr-package}
\title{specifyr: Fast Type Checking}
\description{
Generate functions to quickly check the structure of an object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/specifyr}
  \item Report bugs at \url{https://github.com/EthanSansom/specifyr/issues}
}

}
\keyword{internal}
