% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyr-package.R
\docType{package}
\name{specifyr-package}
\alias{specifyr}
\alias{specifyr-package}
\title{specifyr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EthanSansom/specifyr}
  \item Report bugs at \url{https://github.com/EthanSansom/specifyr/issues}
}

}
\keyword{internal}
